requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
elasticsearch==8.14.0
google-generativeai==0.7.2
python-dotenv==1.0.1